requirements.txt
PySide6
colorama
filetype
tinytag
mutagen
pillow
sounddevice
numpy
pydub
pytubefix
ffpyplayer
yt-dlp
requests
aiohttp
opencv-python
scikit-learn
httpx
mediate
httpx==0.27.2
python-vlc
pyautogui
pywin32
imagehash
